Conditional rendering of <Navbar /> under different Layouts detected in _app.tsx.
Since the _app.tsx layer appears above page components, we need to check if the page components are Home or Store, so that we can render a persistent <Navbar /> component. This persistent <Navbar /> exists between Home or Store.

Should the Page component be something else, like a checkout flow, we do not want the <Navbar /> component to be there, as it will be replaced by something else.