isStoreCartSheet assigned to CartSheets that are opened in the store.
The logic is as follows:

From the Home Page, the ShoppingCart will toggle and show the CartSheet in ANY viewport width.

From the Store Page, the ShoppingCart and CartSheet will toggle and show DEPENDING on the viewport width. If the vw is GREATER than 1185, DISABLE the button and set CartSheet display to none. If it is LESS than 1185, ENABLE the button and set the CartSheet display to flex.

One incredibly important note - You MUST pass a prop that signifies whether or not the the CartSheet comes from a Home or Store component. (_app.tsx, line 83)

This way, we can disable the CartSheet if the vw is greater than 1185 AND it's coming from the Store page. We want this since there will already be an embedded CartOverview for vw>1185px